+++ src/Azure.DevOps.PSModule/Public/Pipeline/Environment/Get-AdoEnvironment.ps1
@@ -102,7 +102,7 @@
 
     process {
         try {
-            $QueryParameters = [system.Collections.Generic.List[string]]::new()
+            $QueryParameters = [List[string]]::new()
 
             if ($Id) {
                 $uri = "$CollectionUri/$ProjectName/_apis/pipelines/environments/$Id"
@@ -133,7 +133,7 @@
                 $continuationToken = $null
 
                 do {
-                    $pagedParams = [System.Collections.Generic.List[string]]::new()
+                    $pagedParams = [List[string]]::new()
 
                     if ($QueryParameters.Count) {
                         $pagedParams.AddRange($QueryParameters)