+++ src/Azure.DevOps.PSModule/Public/ServiceEndpoint/EndPoints/Remove-AdoServiceEndpoint.ps1
@@ -72,7 +72,7 @@
 
     process {
         try {
-            $QueryParameters = [System.Collections.Generic.List[string]]::new()
+            $QueryParameters = [List[string]]::new()
             $QueryParameters.Add("projectIds=$($ProjectIds -join ',')")
 
             if ($Deep.IsPresent) {