+++ src/Azure.DevOps.PSModule/Public/WorkItemTracking/ClassificationNodes/Remove-AdoClassificationNode.ps1
@@ -107,7 +107,7 @@ function Remove-AdoClassificationNode {
 
     process {
         try {
-            $QueryParameters = [System.Collections.Generic.List[string]]::new()
+            $QueryParameters = [List[string]]::new()
 
             if ($ReclassifyId) {
                 $QueryParameters.Add("`$reclassifyId=$ReclassifyId")